requirement.txt
Django==3.2.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
